backend/pyproject.toml
[project]
name = "aix"
version = "0.1.0"
description = "AI Expense Tracker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "ollama>=0.4.7",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "psycopg2>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.0",
]
